ArC - fix handling of disposed parameter qualifiers
- resolves #1839
