+++ independent-projects/arc/processor/src/main/java/io/quarkus/arc/processor/Annotations.java
@@ -17,8 +17,12 @@
 package io.quarkus.arc.processor;
 
 import java.util.Collection;
+import java.util.HashSet;
+import java.util.Set;
 import org.jboss.jandex.AnnotationInstance;
+import org.jboss.jandex.AnnotationTarget.Kind;
 import org.jboss.jandex.DotName;
+import org.jboss.jandex.MethodInfo;
 
 public final class Annotations {
 
@@ -61,6 +65,17 @@ public final class Annotations {
             }
         }
         return false;
+    }
+
+    static Set<AnnotationInstance> getParameterAnnotations(BeanDeployment beanDeployment, MethodInfo method, int position) {
+        Set<AnnotationInstance> annotations = new HashSet<>();
+        for (AnnotationInstance annotation : beanDeployment.getAnnotations(method)) {
+            if (Kind.METHOD_PARAMETER == annotation.target().kind()
+                    && annotation.target().asMethodParameter().position() == position) {
+                annotations.add(annotation);
+            }
+        }
+        return annotations;
     }
 
 }
