+++ independent-projects/arc/processor/src/main/java/io/quarkus/arc/processor/BeanGenerator.java
@@ -1501,7 +1501,7 @@ public class BeanGenerator extends AbstractGenerator {
             annotations = beanDeployment.getAnnotations(field);
         } else {
             MethodInfo method = injectionPoint.getTarget().asMethod();
-            annotations = InjectionPointInfo.getParameterAnnotations(beanDeployment,
+            annotations = Annotations.getParameterAnnotations(beanDeployment,
                     method, injectionPoint.getPosition());
         }
         for (AnnotationInstance annotation : annotations) {
